place the bulk of the `allocat_vars` logic in a standalone generic fn
